Build libraries only if not previously done

+++ wheels/build-common.sh
@@ -64,31 +64,37 @@ function build_libtool {
 function build_dep {
     case $1 in
         'geos')
-            cd /tmp/src
-            curl -f -L -O http://download.osgeo.org/geos/geos-3.6.1.tar.bz2
-            tar jxf geos-3.6.1.tar.bz2
-            cd geos-3.6.1
-            ./configure
-            make -j $NPROC
-            make install
+            if [ ! -f /usr/local/lib/libgeos-3.6.1.so ]; then
+                cd /tmp/src
+                curl -f -L -O http://download.osgeo.org/geos/geos-3.6.1.tar.bz2
+                tar jxf geos-3.6.1.tar.bz2
+                cd geos-3.6.1
+                ./configure
+                make -j $NPROC
+                make install
+            fi
             ;;
         'proj')
-            cd /tmp/src
-            curl -f -L -O http://download.osgeo.org/proj/proj-4.9.3.tar.gz
-            tar xzf proj-4.9.3.tar.gz
-            cd proj-4.9.3
-            ./configure
-            make -j $NPROC
-            make install
+            if [ ! -f /usr/local/lib/libproj.so.12.0.0 ]; then
+                cd /tmp/src
+                curl -f -L -O http://download.osgeo.org/proj/proj-4.9.3.tar.gz
+                tar xzf proj-4.9.3.tar.gz
+                cd proj-4.9.3
+                ./configure
+                make -j $NPROC
+                make install
+            fi
             ;;
         'jasper')
-            cd /tmp/src
-            curl -f -L -O http://download.osgeo.org/gdal/jasper-1.900.1.uuid.tar.gz
-            tar xzf jasper-1.900.1.uuid.tar.gz
-            cd jasper-1.900.1.uuid
-            ./configure --disable-debug --enable-shared
-            make -j $NPROC
-            make install
+            if [ ! -f /usr/local/lib/libjasper.so.1.0.0 ]; then
+                cd /tmp/src
+                curl -f -L -O http://download.osgeo.org/gdal/jasper-1.900.1.uuid.tar.gz
+                tar xzf jasper-1.900.1.uuid.tar.gz
+                cd jasper-1.900.1.uuid
+                ./configure --disable-debug --enable-shared
+                make -j $NPROC
+                make install
+            fi
             ;;
     esac
 }
